+++ modules/client-common/src/internal/interfaces.ts
@@ -21,6 +21,16 @@ export interface IClientWeb3Core {
     abi: ContractInterface,
   ) => Contract & T;
   getDaoFactoryAddress: () => string;
+  getPluginSetupProcessorAddress: () => string;
+  getMultisigRepoAddress: () => string;
+  getAdminRepoAddress: () => string;
+  getAddresslistVotingRepoAddress: () => string;
+  getTokenVotingRepoAddress: () => string;
+  getMultisigSetupAddress: () => string;
+  getAdminSetupAddress: () => string;
+  getAddresslistVotingSetupAddress: () => string;
+  getTokenVotingSetupAddress: () => string;
+  getEnsRegistryAddress: () => string;
   getApproximateGasFee: (estimatedFee: bigint) => Promise<GasFeeEstimation>;
 }
 export interface IClientIpfsCore {