+++ modules/client/src/multisig/internal/client/methods.ts
@@ -59,7 +59,6 @@ import {
   ClientCore,
   EMPTY_PROPOSAL_METADATA_LINK,
   findLog,
-  LIVE_CONTRACTS,
   MULTI_FETCH_TIMEOUT,
   prepareGenericInstallation,
   PrepareInstallationStepValue,
@@ -251,7 +250,7 @@ export class MultisigClientMethods extends ClientCore
     // todo params
     yield* prepareGenericInstallation(this.web3, {
       daoAddressOrEns: params.daoAddressOrEns,
-      pluginRepo: LIVE_CONTRACTS[networkName].multisigRepo,
+      pluginRepo: this.web3.getMultisigRepoAddress(),
       version: params.versionTag,
       installationAbi: INSTALLATION_ABI,
       installationParams: [
@@ -261,6 +260,7 @@ export class MultisigClientMethods extends ClientCore
           params.settings.votingSettings.minApprovals,
         ],
       ],
+      pluginSetupProcessorAddress: this.web3.getPluginSetupProcessorAddress(),
     });
   }
   /**