+++ modules/client/CHANGELOG.md
@@ -19,7 +19,8 @@ TEMPLATE:
 ## [UPCOMING]
 ### Added
 - Block param on `getVotingSettings` and `getMembers` functions to allow for historical data
-## 1.11.0-rc1
+- Support for local chains
+## 1.12.0-rc1
 ### Added
 - Support for baseMainnet network
 ## 1.11.0-rc1